Refactor to use module names as logger names
