+++ nanome/_internal/logging.py
@@ -1,0 +1,101 @@
+import json
+import logging
+import sys
+from dateutil import parser
+from nanome._internal._network import _Packet
+
+
+class LogTypes:
+    """Log Codes as expected by NTS."""
+    DEBUG = 0
+    INFO = 1
+    WARNING = 2
+    ERROR = 3
+
+
+class NTSFormatter(logging.Formatter):
+    """Send NTS json data with specified log level numbers."""
+
+    datefmt = "%Y-%m-%dT%H:%M:%S"
+
+    fmt = {
+        'timestamp': '%(asctime)s',
+        'msg': '%(message)s',
+        'sev': '%(levelname)s',  # Will be manually updated to val from LogType enum.
+    }
+
+    def __init__(self, fmt=None, **kwargs):
+        # Use format saved by class, so no need to pass fmt kwarg
+        fmt = json.dumps(self.fmt)
+        super(NTSFormatter, self).__init__(fmt=fmt, datefmt=self.datefmt, **kwargs)
+
+    def format(self, record):
+        msg = super(NTSFormatter, self).format(record)
+        json_msg = json.loads(msg.replace('\n', '\\n'))
+
+        # Convert timestamp to UTC
+        timestamp = json_msg['timestamp']
+        timestamp_dt = parser.parse(timestamp)
+        json_msg['timestamp'] = timestamp_dt.strftime(self.datefmt)
+
+        # Replace `sev` value with corresponding LogType from enum.
+        level_name = json_msg['sev']
+        enum_val = getattr(LogTypes, level_name)
+        json_msg['sev'] = enum_val
+        updated_msg = json.dumps(json_msg)
+        return updated_msg
+
+
+class NTSLoggingHandler(logging.Handler):
+    """Forward Log messages to NTS."""
+
+    def __init__(self, plugin, *args, **kwargs):
+        super(NTSLoggingHandler, self).__init__(*args, **kwargs)
+        self._plugin = plugin
+        # Unique Identifier for current Nanome session
+        self.formatter = NTSFormatter()
+
+    def handle(self, record):
+        # Use new NTS message format to forward logs.
+        fmted_msg = self.formatter.format(record)
+        packet = _Packet()
+        packet.set(0, _Packet.packet_type_live_logs, 0)
+        packet.write_string(fmted_msg)
+        self._plugin._network.send(packet)
+
+
+class ColorFormatter(logging.Formatter):
+    """Print log outputs in color.
+
+    https://stackoverflow.com/a/56944256
+    """
+
+    grey = "\x1b[0m"
+    yellow = "\x1b[33m"
+    red = "\x1b[91m"
+    bold_red = "\x1b[91m"
+    reset = "\x1b[0m"
+    formats = {}
+
+    def __init__(self, fmt=None, **kwargs):
+        super(ColorFormatter, self).__init__(fmt, **kwargs)
+        self.formats = {
+            logging.DEBUG: self.grey + fmt + self.reset,
+            logging.INFO: self.grey + fmt + self.reset,
+            logging.WARNING: self.yellow + fmt + self.reset,
+            logging.ERROR: self.red + fmt + self.reset,
+            logging.CRITICAL: self.bold_red + fmt + self.reset
+        }
+
+    def format(self, record):
+        log_fmt = self.formats.get(record.levelno)
+        formatter = logging.Formatter(log_fmt)
+        if self.supports_color():
+            output = formatter.format(record)
+        else:
+            output = super(ColorFormatter, self).format(record)
+        return output
+
+    @staticmethod
+    def supports_color():
+        return not sys.platform == 'win32' or not sys.stdout.isatty()
