+++ nanome/util/logs.py
@@ -1,6 +1,8 @@
 import sys
 import functools
+import inspect
 from .enum import IntEnum, auto
+import logging
 
 
 class Logs(object):
@@ -52,8 +54,10 @@ class Logs(object):
         :param args: Variable length argument list
         :type args: Anything printable
         """
-        log_type = cls.LogType.error.name
-        cls._log(log_type, *args)
+        module = cls.caller_name()
+        logger = logging.getLogger(module)
+        msg = ' '.join(map(str, args))
+        logger.error(msg)
 
     @classmethod
     def warning(cls, *args):
@@ -63,8 +67,10 @@ class Logs(object):
         :param args: Variable length argument list
         :type args: Anything printable
         """
-        log_type = cls.LogType.warning.name
-        cls._log(log_type, *args)
+        module = cls.caller_name()
+        logger = logging.getLogger(module)
+        msg = ' '.join(map(str, args))
+        logger.warning(msg)
 
     @classmethod
     def message(cls, *args):
@@ -74,8 +80,10 @@ class Logs(object):
         :param args: Variable length argument list
         :type args: Anything printable
         """
-        log_type = cls.LogType.info.name
-        cls._log(log_type, *args)
+        module = cls.caller_name()
+        logger = logging.getLogger(module)
+        msg = ' '.join(map(str, args))
+        logger.info(msg)
 
     @classmethod
     def debug(cls, *args):
@@ -86,12 +94,10 @@ class Logs(object):
         :param args: Variable length argument list
         :type args: Anything printable
         """
-        log_type = cls.LogType.debug.name
-        if cls._verbose is None:
-            Logs.warning("Debug used before plugin start.")
-            cls._log(log_type, *args)
-        elif cls._verbose is True:
-            cls._log(log_type, *args)
+        module = cls.caller_name()
+        logger = logging.getLogger(module)
+        msg = ' '.join(map(str, args))
+        logger.debug(msg)
 
     @staticmethod
     def deprecated(new_func=None, msg=""):
@@ -109,3 +115,42 @@ class Logs(object):
             wrapper.used = False
             return wrapper
         return deprecated_decorator
+
+    @staticmethod
+    def caller_name(skip=2):
+        """Get a name of a caller in the format module.class.method
+
+        `skip` specifies how many levels of stack to skip while getting caller
+        name. skip=1 means "who calls me", skip=2 "who calls my caller" etc.
+
+        An empty string is returned if skipped levels exceed stack height
+
+        https://stackoverflow.com/questions/2654113/how-to-get-the-callers-method-name-in-the-called-method
+        """
+        stack = inspect.stack()
+        start = 0 + skip
+        if len(stack) < start + 1:
+            return ''
+        parentframe = stack[start][0]
+
+        name = []
+        module = inspect.getmodule(parentframe)
+        # `modname` can be None when frame is executed directly in console
+        # TODO(techtonik): consider using __main__
+        if module:
+            name.append(module.__name__)
+        # detect classname
+        if 'self' in parentframe.f_locals:
+            # I don't know any way to detect call from the object method
+            # XXX: there seems to be no way to detect static method call - it will
+            #      be just a function call
+            name.append(parentframe.f_locals['self'].__class__.__name__)
+        codename = parentframe.f_code.co_name
+        if codename != '<module>':  # top level usually
+            name.append(codename)  # function or a method
+
+        # Avoid circular refs and frame leaks
+        #  https://docs.python.org/2.7/library/inspect.html#the-interpreter-stack
+        del parentframe, stack
+
+        return ".".join(name)
